compile fixes and add some things

+++ src/compression/ZstdCompressor.cpp
@@ -74,7 +74,7 @@ CompressorResult<> ZstdCompressor::compress(
     if (ZSTD_isError(dsize)) {
         auto ec = ZSTD_getErrorCode(dsize);
         auto msg = ZSTD_getErrorName(ec);
-        log::warn("ZstdCompressor: Decompression failed: {} ({})", msg, ec);
+        log::warn("ZstdCompressor: Decompression failed: {} ({})", msg, (int) ec);
         return Err(CompressorError::CompressionFailed);
     }
